requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
reportlab==4.0.7
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
